Change styles and remove old about  component
