+++ src/app/app.component.html
@@ -1,7 +1,10 @@
 <app-navbar></app-navbar>
 <main>
-  <section id="home">
-    <app-home></app-home>
+  <section id="about">
+    <app-about></app-about>
+  </section>
+  <section id="stack">
+    <app-stack></app-stack>
   </section>
   <section id="work">
     <app-work></app-work>
