+++ src/app/app.component.scss
@@ -17,7 +17,7 @@ main {
   }
 }
 
-#home {
+#about {
   padding-top: 8rem;
 
   @include media-query-min('m') {
