+++ src/app/app.routes.ts
@@ -1,12 +1,10 @@
 import { Routes } from '@angular/router';
-import { HomeComponent } from './features/home/home.component';
-import { AboutComponent } from './features/about/about.component';
 import { WorkComponent } from './features/work/work.component';
 import { PortfolioComponent } from './features/portfolio/portfolio.component';
+import { AboutComponent } from './features/about/about.component';
 
 export const routes: Routes = [
-  { path: '', component: HomeComponent },
-  { path: 'home', component: HomeComponent },
+  { path: '', component: AboutComponent },
   { path: 'about', component: AboutComponent },
   { path: 'work', component: WorkComponent },
   { path: 'portfolio', component: PortfolioComponent },
