--- src/app/features/home/home.component.html
@@ -1,21 +1,0 @@
-<div class="home-container">
-  <div class="image-container">
-    <img src="../../../assets/images/image01.jpeg" alt="Profile Picture" />
-  </div>
-
-  <div class="content-container">
-    <h1 class="title">{{ 'myName' | translate }},</h1>
-    <h2 class="title">
-      <span [innerHTML]="typedText"></span>
-    </h2>
-
-    <p class="introduction">
-      {{ 'homePage.introduction' | translate }}
-    </p>
-    <div class="skills-container">
-      <div class="skill" *ngFor="let skill of skills">
-        {{ skill.name }}
-      </div>
-    </div>
-  </div>
-</div>
