--- src/app/features/home/home.component.scss
@@ -1,125 +1,0 @@
-@use '../../../../src/styles' as *;
-@use '../../../assets/shared-styles/mediaQueries.scss' as *;
-@use '../../../assets/shared-styles/colors.scss' as *;
-@use '../../../assets/shared-styles/text.scss' as *;
-
-.home-container {
-  display: grid;
-  grid-template-columns: 1fr;
-  gap: 20px;
-}
-
-.image-container {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  margin-bottom: 3rem;
-}
-
-.image-container img {
-  width: 14rem;
-  border-radius: 50%;
-
-  &:hover {
-    box-shadow: 0px 0px 10px 2px #ff6347;
-  }
-}
-
-.content-container {
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-}
-
-.title {
-  margin: 0;
-  line-height: 1.1;
-  letter-spacing: 1px;
-  color: #333;
-  font-weight: bold;
-}
-
-.title:first-child {
-  color: #ff6347;
-  font-size: $font-size-xxlarge;
-}
-
-.title:nth-child(2) {
-  color: #4682b4;
-  font-size: $font-size-large;
-}
-
-.title span {
-  animation: blink-cursor 0.7s steps(2, start) infinite;
-}
-
-@keyframes blink-cursor {
-  50% {
-    border-color: transparent;
-  }
-}
-
-.introduction {
-  font-size: $font-size-small;
-  margin-bottom: 20px;
-  text-align: left;
-  line-height: 1.8;
-}
-
-.skills-container {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 15px;
-  margin-top: 1rem;
-  font-family: $font-family-skills;
-}
-
-.skill {
-  padding: 0.5rem 1.5rem;
-  font-size: $font-size-medium;
-  border: 2px solid #007acc;
-  border-radius: 30px;
-  transition: all 0.3s ease-in-out;
-  box-shadow: 0 4px 6px rgba(0, 0, 0, 0.1);
-
-  &:hover {
-    background-color: #007acc;
-    color: var(--text-color-primary);
-    box-shadow: 0 6px 8px rgba(0, 0, 0, 0.15);
-    transform: translateY(-3px);
-  }
-
-  &:active {
-    transform: translateY(1px);
-  }
-
-  &:focus {
-    outline: 3px solid #005f99;
-  }
-}
-
-@include media-query-min('m') {
-  .home-container {
-    grid-template-columns: 1fr 2fr;
-  }
-
-  .image-container {
-    display: contents;
-  }
-
-  .image-container img {
-    width: 18rem;
-  }
-
-  .content-container {
-    padding-left: 20px;
-  }
-
-  .title:first-child {
-    font-size: $font-size-xxxlarge;
-  }
-
-  .title:nth-child(2) {
-    font-size: $font-size-xxlarge;
-  }
-}
