+++ src/app/features/stack/stack.component.html
@@ -1,0 +1,6 @@
+<h2 class="stack-title">{{ 'myStack' | translate }}</h2>
+<div class="skills-container">
+  <div class="skill" *ngFor="let skill of skills">
+    {{ skill.name }}
+  </div>
+</div>
