+++ src/app/features/work/work.component.html
@@ -24,48 +24,25 @@
 <div class="job-details-main" *ngIf="selectedJobIndex !== null">
   <div *ngFor="let role of jobs[selectedJobIndex].roles; let j = index">
     <div class="job-details-container">
+      <div class="job-logo">
+        <a [href]="jobs[selectedJobIndex].website" target="_blank">
+          <img
+            [src]="jobs[selectedJobIndex].logo"
+            alt="{{ jobs[selectedJobIndex].company }} Logo"
+            class="logo"
+          />
+        </a>
+      </div>
       <div class="job-details">
         <div class="job-content">
           <div class="column-1">
-            <ul>
-              <li>
-                <strong>{{ 'company' | translate }}:</strong>
-                {{ jobs[selectedJobIndex].company }}
-                <span *ngIf="jobs[selectedJobIndex].newName"
-                  >({{ jobs[selectedJobIndex].newName }})</span
-                >
-              </li>
-              <li>
-                <strong>{{ 'role' | translate }}:</strong> {{ role.title }}
-              </li>
-              <li>
-                <strong>{{ 'dates' | translate }}:</strong>
-                {{ role.startDate }} -
-                {{ role.endDate || ('present' | translate) }}
-              </li>
-              <li>
-                <strong>{{ 'technologies' | translate }}:</strong>
-                {{ role.technologies.join(', ') }}
-              </li>
-            </ul>
+            <strong class="job-company">{{
+              jobs[selectedJobIndex].company
+            }}</strong>
+            <div class="job-role">{{ role.title }}</div>
+            <div class="job-stack">{{ role.technologies.join(', ') }}</div>
+            <div class="job-description">{{ role.description }}</div>
           </div>
-          <div class="column-2">
-            <div class="job-logo">
-              <a [href]="jobs[selectedJobIndex].website" target="_blank">
-                <img
-                  [src]="jobs[selectedJobIndex].logo"
-                  alt="{{ jobs[selectedJobIndex].company }} Logo"
-                  class="logo"
-                />
-              </a>
-            </div>
-          </div>
-        </div>
-        <div class="job-description" *ngIf="role.description">
-          <p>
-            <strong>{{ 'description' | translate }}:</strong>
-            {{ role.description }}
-          </p>
         </div>
       </div>
     </div>
